Redesigned Vector class to be a subclass of Matrix with direct references to the underlying data for x/y/z members.  This needs to be extended to work with 4D vectors and unit tests need to be written for basic ops as well as vector access.  Additionally, further constructors should be added to the Vector class to allow construction from initializer lists, etc.

+++ Utils/build_units.py
@@ -51,7 +51,7 @@ unit_template = Template(
         static inline Q_${symbol}<long double> operator "" _${symbol}_ (long double _v) {
             return Q_${symbol}<long double>(_v); }
         static inline Q_${symbol}<long double> operator "" _${symbol}_ (unsigned long long int _v) {
-            return Q_${symbol}<long double>(_v); }
+            return Q_${symbol}<long double>(static_cast<long double>(_v)); }
     }
 """)
